code clean up

weight loading cleanup


rename oai -> openaimoe for HF compat


format


finished rebase
