+++ vllm/model_executor/layers/quantization/mxfp4.py
@@ -129,7 +129,13 @@ class Mxfp4MoEMethod(FusedMoEMethodBase):
         e_score_correction_bias: Optional[torch.Tensor] = None,
         apply_router_weight_on_input: bool = False,
         activation: str = "silu",
+        enable_eplb: bool = False,
+        expert_load_view: Optional[torch.Tensor] = None,
+        logical_to_physical_map: Optional[torch.Tensor] = None,
+        logical_replica_count: Optional[torch.Tensor] = None,
     ) -> torch.Tensor:
+        if enable_eplb:
+            raise NotImplementedError("EPLB is not supported for mxfp4")
 
         return triton_kernel_moe_forward(
             hidden_states=x,
