+++ vllm/model_executor/models/registry.py
@@ -138,7 +138,7 @@ _TEXT_GENERATION_MODELS = {
     "TeleFLMForCausalLM": ("teleflm", "TeleFLMForCausalLM"),
     "XverseForCausalLM": ("llama", "LlamaForCausalLM"),
     "Zamba2ForCausalLM": ("zamba2", "Zamba2ForCausalLM"),
-    "OAIForCausalLM": ("oai", "OAIForCausalLM"),
+    "OpenAIMoeForCausalLM": ("openai_moe", "OpenAIMoeForCausalLM"),
     # [Encoder-decoder]
     "BartModel": ("bart", "BartForConditionalGeneration"),
     "BartForConditionalGeneration": ("bart", "BartForConditionalGeneration"),
